Add files via upload
The Udacity--Project-Investigate-TMDB-Movies-Dataset project is a data analysis project that uses the Python libraries NumPy, pandas, and Matplotlib to analyze a dataset of about 10,000 movies collected from TMDb. The goal of the project is to analyze insights about movie features to help the movie industry earn more profits.

The project requires the installation of Python and the following libraries: pandas, NumPy, Matplotlib, and csv. It is recommended to install Anaconda, which comes with all of the necessary packages, as well as IPython notebook.

The project teaches the following skills:

The steps involved in a typical data analysis process
How to pose questions that can be answered with a given dataset and then answer those questions
How to investigate problems in a dataset and wrangle the data into a format you can use
How to communicate the results of your analysis
How to use vectorized operations in NumPy and pandas to speed up your data analysis code
How to use pandas' Series and DataFrame objects
How to use Matplotlib to produce plots showing your findings